+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Move seek offset ownership to the coordinator to allow Virtual Partitions further development.
 - Improve client shutdown in specs.
 - Do not reset client on network issue and rely on `librdkafka` to do so.
+- Allow for nameless (anonymous) subscription groups (#1033)
 
 ## 2.0.8 (2022-09-19)
 - [Breaking change] Rename Virtual Partitions `concurrency` to `max_partitions` to avoid confusion  (#1023).